+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next release
 
 ### Bug fixes and improvements
+- server: fix parsing remote relationship json definition from 1.x server catalog on migration (fix #7906)
 
 ## v2.6.0-beta.1